backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dlib==20.0.0
dnspython==2.7.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
gunicorn==23.0.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
packaging==25.0
pillow==11.3.0
pymongo==4.13.2
python-dotenv==1.1.1
Werkzeug==3.0.1
zipp==3.23.0
